(feat): added Details API
